+++ src/main/java/com/example/recruit_page_wwy/reply/ReplyRequest.java
@@ -7,22 +7,6 @@ import lombok.NoArgsConstructor;
 
 public class ReplyRequest {
 
-    @Data
-    public static class DTO {
-        private int id;
-        private Board board;
-        private User user;
-        private String content;
-        private String createdAt;
-
-        public DTO(Reply reply, Board board, User sessionUser) {
-            this.id = reply.getId();
-            this.board = board;
-            this.user = sessionUser;
-            this.content = reply.getContent();
-            this.createdAt = reply.getCreatedAt().toString();
-        }
-    }
 
     @Data
     @NoArgsConstructor
