+++ src/main/java/com/example/recruit_page_wwy/reply/ReplyService.java
@@ -19,17 +19,16 @@ public class ReplyService {
 
     // TODO : Reply가 아니라 DTO에 담아서 반환
     @Transactional
-    public ReplyRequest.DTO replySave(ReplyRequest.SaveDTO saveDTO, User sessionUser) {
+    public ReplyResponse.DTO replySave(ReplyRequest.SaveDTO saveDTO, User sessionUser) {
         if (saveDTO.getContent().isBlank()) throw new ExceptionApi400("댓글 내용을 입력하세요.");
         Board board = boardRepository.findById(saveDTO.getBoardId());
         if (board == null) {
             throw new ExceptionApi404("게시글을 찾을 수 없습니다. id = " + saveDTO.getBoardId());
         }
 
-        Reply reply = saveDTO.toEntity(sessionUser);
+        Reply replyPS = replyRepository.replySave(saveDTO.toEntity(sessionUser));
 
-        Reply replyPS = replyRepository.replySave(reply);
-        return new ReplyRequest.DTO(replyPS, board, sessionUser);
+        return new ReplyResponse.DTO(replyPS);
     }
 
     // TODO : 예외 처리
